requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
pandas>=1.5.0
publicdotcom-py>=1.0.0
requests>=2.28.0
